Side effects: replace access_mode with (currently trivial) side_effect_order; split side_effect.h
