+++ include/types.h
@@ -55,3 +55,11 @@ MAKE_PHANTOM_TYPE(command_id, size_t)
 MAKE_PHANTOM_TYPE(collective_group_id, size_t)
 MAKE_PHANTOM_TYPE(reduction_id, size_t)
 MAKE_PHANTOM_TYPE(host_object_id, size_t)
+
+
+// declared in this header for include-dependency reasons
+namespace celerity::experimental {
+
+enum class side_effect_order { sequential };
+
+}
