+++ src/graph_generator.cc
@@ -483,25 +483,15 @@ namespace detail {
 
 			for(const auto& side_effect : tsk->get_side_effect_map()) {
 				auto [hoid, mode] = side_effect;
-				const auto is_producer = mode == access_mode::write || mode == access_mode::read_write;
-				const auto is_consumer = mode == access_mode::read || mode == access_mode::read_write;
-				if(is_producer) {
-					if(const auto last_effect = nd.host_object_last_effects.find(hoid); last_effect != nd.host_object_last_effects.end()) {
-						cdag.add_dependency(cmd, cdag.get(last_effect->second), dependency_kind::ANTI_DEP);
-					}
-				}
-				if(is_consumer) {
-					if(const auto last_producer = nd.host_object_last_producers.find(hoid); last_producer != nd.host_object_last_producers.end()) {
-						cdag.add_dependency(cmd, cdag.get(last_producer->second), dependency_kind::TRUE_DEP);
-					}
+				if(const auto last_effect = nd.host_object_last_effects.find(hoid); last_effect != nd.host_object_last_effects.end()) {
+					cdag.add_dependency(cmd, cdag.get(last_effect->second), dependency_kind::TRUE_DEP);
 				}
 
 				// Simplification: If there are multiple chunks per node, we generate true-dependencies between them in an arbitrary order, when all we really
 				// need is mutual exclusion (i.e. a bi-directional pseudo-dependency).
 				nd.host_object_last_effects.insert_or_assign(hoid, cmd->get_cid());
-				if(is_producer) { nd.host_object_last_producers.insert_or_assign(hoid, cmd->get_cid()); }
 
-				cmd->debug_label += fmt::format("{} host-object {}\n", detail::access::mode_traits::name(mode), hoid);
+				cmd->debug_label += fmt::format("host-object {}\n", hoid);
 			}
 		}
 	}
@@ -533,9 +523,6 @@ namespace detail {
 					cid = std::max(prev_hid, cid);
 				}
 				for(auto& [cgid, cid] : this_node_data.host_object_last_effects) {
-					cid = std::max(prev_hid, cid);
-				}
-				for(auto& [cgid, cid] : this_node_data.host_object_last_producers) {
 					cid = std::max(prev_hid, cid);
 				}
 				// update lowest previous horizon id (for later command deletion)
