+++ test/runtime_tests.cc
@@ -2434,9 +2434,9 @@ namespace detail {
 		experimental::host_object void_ho;
 
 		q.submit([=](handler& cgh) {
-			experimental::side_effect append_owned{owned_ho, cgh, celerity::write_only};
-			experimental::side_effect append_ref{ref_ho, cgh, celerity::write_only};
-			experimental::side_effect track_void{void_ho, cgh, celerity::write_only};
+			experimental::side_effect append_owned{owned_ho, cgh};
+			experimental::side_effect append_ref{ref_ho, cgh};
+			experimental::side_effect track_void{void_ho, cgh};
 			cgh.host_task(on_master_node, [=] {
 				(*append_owned).push_back(1);
 				(*append_ref).push_back(1);
@@ -2454,7 +2454,7 @@ namespace detail {
 		});
 
 		q.submit([=](handler& cgh) {
-			experimental::side_effect check_owned{owned_ho, cgh, celerity::read_only};
+			experimental::side_effect check_owned{owned_ho, cgh};
 			cgh.host_task(on_master_node, [=] { CHECK(*check_owned == std::vector{1, 2}); });
 		});
